Add dashboard components including hero screen details, cards list view, custom hero icon, floating action button, and header
