+++ lib/core/routes/app_routes.dart
@@ -2,7 +2,9 @@
 
 import 'package:az_task/core/routes/base_route.dart';
 import 'package:az_task/features/authentication/ui/views/login_screen.dart';
+import 'package:az_task/features/dashboard/data/card_model.dart';
 import 'package:az_task/features/dashboard/ui/views/dashboard_screen.dart';
+import 'package:az_task/features/dashboard/ui/views/hero_screen_details.dart';
 import 'package:az_task/features/onboarding_screen/ui/views/last_view.dart';
 import 'package:az_task/features/onboarding_screen/ui/views/onboarding_view.dart';
 import 'package:az_task/features/splash_screen/ui/views/splash_view.dart';
@@ -15,6 +17,7 @@ class AppRoutes {
   static const loginScreen = '/loginScreen';
   static const dashboardScreen = '/dashboardScreen';
   static const lastView = '/lastview';
+  static const screenDetailsHero = '/screenDetailsHero';
 
   static Route<void>? onGenerateRoute(RouteSettings settings) {
     final args = settings.arguments;
@@ -29,6 +32,10 @@ class AppRoutes {
         return BaseRoute(page: LoginScreen());
       case dashboardScreen:
         return BaseRoute(page: const DashboardScreen());  
+      case screenDetailsHero:
+        return BaseRoute(page:  HeroScreenDetails(
+          cardModel: args as CardModel,
+        ));  
       default:
         return null;
     }
