+++ lib/features/dashboard/ui/widgets/dashboard_header.dart
@@ -1,0 +1,34 @@
+
+import 'package:az_task/core/common/widgets/spaces.dart';
+import 'package:az_task/core/common/widgets/theme_toggle_switch.dart';
+import 'package:az_task/core/images/assets_images.dart';
+import 'package:flutter/material.dart';
+
+class DashboardHeader extends StatelessWidget {
+  const DashboardHeader({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+      children: [
+        Row(
+          children: [
+            CircleAvatar(
+              backgroundImage: AssetImage(AssetsImages.person),
+              radius: 20,
+            ),
+            horizontalSpace(20),
+            Text(
+              "Welcome to AZ",
+              style: Theme.of(
+                context,
+              ).textTheme.titleMedium?.copyWith(fontWeight: FontWeight.bold),
+            ),
+          ],
+        ),
+        ThemeToggleSwitch(),
+      ],
+    );
+  }
+}
